LLM_Vendor Tool Data/Cursor/Cursor_Unittests_String_Sorted.py
import pytest

def test_ginortS():
    assert ginortS('Sorting1234') == 'ginortS1324'
    assert ginortS('Test1234') == 'estT1324'
    assert ginortS('HelloWorld1234') == 'deHllloorW1324'
    assert ginortS('') == ''