LLM_Vendor Tool Data/pyproject.toml
[tool.poetry]
name = "OpenCiviWiki"
version = "1.0"
description = "Building a Better Democracy for the Internet Age"
authors = ["Brylie Christopher Oxley <brylie@amble.fi>"]
license = "AGPL"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1"
django-cors-headers = "^3.13.0"
django-extensions = "^3.2.0"
djangorestframework = "^3.13.1"
django-notifications-hq = "^1.7.0"
django-storages = "^1.13.1"
whitenoise = "^6.2.0"
psycopg2-binary = "^2.9.3"
django-taggit = "^3.0.0"
Pillow = "^9.2.0"
requests = "^2.28.1"
django-browser-reload = "^1.7.0"
[tool.poetry.dev-dependencies]
black = "^22.6.0"
coverage = "^6.4.4"
flake8 = "^5.0.4"

[tool.poetry.group.dev.dependencies]
pytest-django = "^4.5.2"
ipython = "^8.5.0"
django-debug-toolbar = "^3.7.0"
django-linear-migrations = "^2.5.1"
pre-commit = "^2.20.0"
factory-boy = "^3.2.1"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
